requirements.txt
kivy==2.2.1
kivy_garden.mapview
requests
